refactor(t2): extract MemoryStore from monolith (nexus-vx3c)
RDR-063 Phase 1 step 2 — extract all memory-domain state and methods
out of T2Database into a dedicated MemoryStore class living in
nexus.db.t2.memory_store. T2Database is now a thin facade that
constructs SharedConnection(conn,lock), then MemoryStore(shared), and
delegates every memory call to self.memory.

Moved from t2/__init__.py → t2/memory_store.py:
- Helpers: _FTS5_SPECIAL, _sanitize_fts5, _FTS_REBUILD_SQL,
  _MEMORY_SCHEMA_SQL (the memory + memory_fts portion of the old
  _SCHEMA_SQL), _COLUMNS, _read_session_id (import binding),
  AccessPolicy (Literal type alias)
- Migrations: _migrate_fts_if_needed, _migrate_access_tracking_if_needed
  (now lock-naive methods on MemoryStore — caller holds the lock)
- Class: MemoryStore with _STOPWORDS class attr and 12 public methods:
  put, get, search, list_entries, get_projects_with_prefix,
  search_glob, search_by_tag, get_all, delete,
  find_overlapping_memories, merge_memories, flag_stale_memories,
  plus a new MemoryStore.expire() that returns the deleted row IDs
  so the facade can aggregate metrics across memory + telemetry.

Facade changes (t2/__init__.py):
- _SCHEMA_SQL renamed to _RESIDUAL_SCHEMA_SQL — only plans + topics
  schema; memory schema is now owned by MemoryStore
- T2Database.__init__ creates a SharedConnection wrapping its own
  conn/lock and instantiates self.memory: MemoryStore
- _init_schema runs memory.init_schema_unlocked() then the residual
  DDL under one lock acquisition, then runs all 6 migrations under
  the same _migrated_lock guard (memory ones via self.memory)
- All 12 memory methods become one-line delegates to self.memory.*
- T2Database.expire() composes self.memory.expire() (memory rows) +
  self.expire_relevance_log() and emits the same expire_complete
  structured log with memory_deleted + relevance_log_deleted fields,
  preserving the contract enforced by test_structlog_events.py
- _sanitize_fts5 re-exported at facade level so
  nexus.catalog.catalog_db and tests/test_t2.py keep their existing
  import paths working unchanged

Lock semantics preserved:
- Public memory methods on MemoryStore acquire self._lock themselves
  (the same lock object as the facade's, in Phase 1)
- init_schema_unlocked + migration helpers are lock-naive — the
  facade's _init_schema holds the lock once for the whole
  schema/migration sequence, identical to the pre-split race-free
  guard semantics. Verified by 5/5 stable runs of
  test_migration_guard_concurrent_threads /
  test_migration_guard_path_normalization /
  test_migration_guard_sequential_construction.

Test update:
- tests/test_memory.py::test_t2_uses_session_module_for_session_id
  now patches nexus.db.t2.memory_store._read_session_id (the new
  binding location) instead of nexus.db.t2._read_session_id. This
  is the only test internal that probed the old location; all
  production callers continue to import unchanged.

Tests: 469 passed across all 17 T2-touching files; zero regressions.

Refs: docs/rdr/rdr-063-t2-domain-split.md
Bead: nexus-vx3c (closes), epic nexus-4y8d
